+++ src/core/encryption/__tests__/utils.ts
@@ -1,0 +1,6 @@
+export function fromHex(str: string) {
+	return new Uint8Array(Buffer.from(str, 'hex'));
+}
+export function toHex(buf: Uint8Array) {
+	return Buffer.from(buf).toString('hex').toUpperCase();
+}
